Integrated Calculator into the console shell
Theres still some errors and bugs. Also need to put stage controls and awareness.

+++ KSPCraftParser/PartDataReaderLibrary/Calculators/ThrustCalculator.cs
@@ -91,7 +91,7 @@ namespace PartDataReaderLibrary.Calculators
 		{
 			foreach (var part in partList)
 			{
-				TotalDryMass += part.SearchValue(name);
+				TotalDryMass += (double)part.SearchValue(name);
 			}
 		}
 		#endregion
